backend/requirements.txt
# Neural Roots AI - Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.6.1

# Settings & Configuration
pydantic==2.5.3
pydantic-settings==2.1.0

# Twilio (WhatsApp)
twilio==8.10.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# AI - Gemini API
google-genai>=1.0.0

# AI/ML (for future use)
# tensorflow==2.15.0
# numpy==1.26.3
# pillow==10.2.0

# Utilities
python-dotenv==1.0.0
